add github releases backend (#62)
* go.mod: pull k-sigs/release-sdk

Signed-off-by: Adolfo Garcia Veytia (puerco) <puerco@chainguard.dev>

* Add github release storage backend

This commit adds support to tejolote to read artifacts from github releases.
The url scheme is: github://org/repo/tag

In order to build its inventory, Tejolote will download all the assets in the
release and checksum them.

Signed-off-by: Adolfo Garcia Veytia (puerco) <puerco@chainguard.dev>

* Add github release backend test

Signed-off-by: Adolfo Garcia Veytia (puerco) <puerco@chainguard.dev>

* Add options, ignore extensions

This commit adds an options struct to the github release backend. It
now supports ignoring files and by default it will ignore files with .pem
.sig and .cert extensions, thus not adding them to the inventory of subjects.

Signed-off-by: Adolfo Garcia Veytia (puerco) <puerco@chainguard.dev>

* Storage: Register github:// schema

Signed-off-by: Adolfo Garcia Veytia (puerco) <puerco@chainguard.dev>

---------

Signed-off-by: Adolfo Garcia Veytia (puerco) <puerco@chainguard.dev>